Se hace primer paso para migrar a EC2
